requirements.txt
pytesseract == 0.3.8
pdf2image == 1.16.0
Pillow == 8.1.2
fpdf == 1.7.2
tkinter
PyPDF2 == 1.26.0